:recycle: Refactor WSI `open` - Move `WSIReader.open` and all `try_*` functions to `factory.py` - WSIReader then imports lazily `open_wsi` from `factory.py`. - Move all `is_*` functions to `detection.py` - Move `WSIReaderParams` to `types.py`
