+++ tiatoolbox/wsicore/__init__.py
@@ -1,8 +1,5 @@
 """Package to read whole slide images."""
 
-from numbers import Number
-from typing import TypedDict
-
 from tiatoolbox.wsicore import metadata, wsimeta, wsireader
 
 from .wsimeta import WSIMeta
@@ -15,11 +12,4 @@ __all__ = [
     "WSIReaderParams",
 ]
 
-
-class WSIReaderParams(TypedDict, total=False):
-    """Parameters for reading whole slide images."""
-
-    meta: WSIMeta | None
-    mpp: tuple[Number, Number] | Number
-    power: Number
-    storage_options: dict  # For FsspecStore
+from .wsireader.types import WSIReaderParams
