+++ tiatoolbox/wsicore/wsireader/__init__.py
@@ -21,12 +21,8 @@ from .base import (
     _handle_tiff_wsi,
     _handle_virtual_wsi,
     fix_mangled_url_by_pathlib,
-    is_dicom,
-    is_ngff,
-    is_tiled_tiff,
-    is_url,
-    is_zarr,
 )
+from .detection import is_dicom, is_ngff, is_tiled_tiff, is_url, is_zarr
 
 __all__ = [
     "AnnotationStoreReader",
